requirements.txt
bcrypt==3.2.0
cffi==1.14.5
fastapi==0.63.0
greenlet==1.0.0
pycparser==2.20
pydantic==1.8.1
PyMySQL==1.0.2
python-dotenv==0.17.0
six==1.15.0
SQLAlchemy==1.4.5
starlette==0.13.6
typing-extensions==3.7.4.3
